Remove unnecessary casts in RetryUtils

+++ langchain4j-core/src/main/java/dev/langchain4j/internal/RetryUtils.java
@@ -151,7 +151,7 @@ public final class RetryUtils {
          * @return The raw delay in milliseconds.
          */
         public double rawDelayMs(int retry) {
-            return ((double) delayMillis) * Math.pow(backoffExp, retry);
+            return delayMillis * Math.pow(backoffExp, retry);
         }
 
         /**
